# openapi.json
{
  "components": {
    "schemas": {
...
      "ContentType": {
        "description": "Engine category that supports backend overrides.\n\nCurrently only ASR and FA have multiple engine backends.\nOther inference tasks (morphosyntax, utseg, translate, coref)\nalways use their single built-in engine.\nMIME-like content discriminator for file results.",
        "enum": [
          "chat",
          "csv",
          "text",
          "json"
        ],
        "type": "string"
      },
...
      "DisplayPath": {
        "description": "Display path for a file within a job: either a bare basename\n(`\"sample.cha\"`) for single-file input or a relative forward-slash path\n(`\"PWA/TYO_a1.cha\"`) for directory input with subdirectories.\n\nBackslashes are normalized to forward slashes on construction so the value\nis platform-independent regardless of whether the CLI ran on Windows.\n\nThis type replaces the former `FileName` which incorrectly rejected path\nseparators during deserialization even though the system routinely carries\nrelative paths.",
        "type": "string"
      },
...
      "ErrorResponse": {
        "description": "Standard error response body, matching FastAPI's `HTTPException` format.\n\nAll non-2xx responses from the server use this shape so that clients\ncan parse errors uniformly. The one exception is\n[`ServerError::JobConflict`](crate::error::ServerError::JobConflict),\nwhich nests a `conflicts` array inside `detail`.",
        "properties": {
          "detail": {
            "description": "Human-readable description of what went wrong. For validation\nerrors this includes the specific field or constraint that failed.",
            "type": "string"
          }
        },
        "required": [
          "detail"
        ],
        "type": "object"
      },
...
      "ExecutionPlan": {
        "description": "Execution plan describing where and how a job is processed.\n\nAttached to `Job` and projected into `JobInfo` for API consumers.\nThe `stage` field tracks progress through the staged-remote lifecycle;\nfor local and explicit-remote jobs it is always [`ExecutionStage::Done`]\nonce the job finishes.",
        "properties": {
          "execution_host": {
            "description": "Hostname of the machine executing the job (e.g. `\"server-01\"`).",
            "type": "string"
          },
          "mode": {
            "$ref": "#/components/schemas/ExecutionMode",
            "description": "How the job is being executed."
          },
          "remote_job_id": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/JobId",
                "description": "For staged-remote jobs: the job ID on the remote server.\n`None` until the remote job is submitted."
              },
              {
                "type": "null"
              }
            ]
          },
          "stage": {
            "$ref": "#/components/schemas/ExecutionStage",
            "description": "Current lifecycle stage of the execution plan."
          }
        },
        "required": [
          "mode",
          "execution_host",
          "stage"
        ],
        "type": "object"
      },
...
      "FileResult": {
        "description": "Result for a single processed file.",
        "properties": {
          "content": {
            "description": "Processed file content (CHAT text or CSV).  Empty string when\n`error` is `Some`.",
            "type": "string"
          },
          "content_type": {
            "$ref": "#/components/schemas/ContentType",
            "description": "MIME-like content discriminator: `\"chat\"` for CHAT files (default),\n`\"csv\"` for tabular output (e.g. opensmile features)."
          },
          "error": {
            "description": "Human-readable error message if this file failed processing.\n`None` for successfully processed files.",
            "type": [
              "string",
              "null"
            ]
          },
          "filename": {
            "$ref": "#/components/schemas/DisplayPath",
            "description": "Display path for this file: a bare basename (`\"sample.cha\"`) or a\nrelative forward-slash path (`\"PWA/TYO_a1.cha\"`) for directory input.\nBackslashes are normalized to forward slashes on construction."
          },
          "provenance": {
            "$ref": "#/components/schemas/FileProvenance",
            "description": "Processing provenance read from the output file, as a typed state:\nthe parsed stamps, a stamp of ours that did not parse, or nothing read\n(non-CHAT output, or a file that failed)."
          }
        },
        "required": [
          "filename",
          "provenance"
        ],
        "type": "object"
      },
...
      "JobControlPlaneInfo": {
        "description": "Backend-owned orchestration metadata for a job projection.",
        "properties": {
          "backend": {
            "$ref": "#/components/schemas/JobControlPlaneBackendKind",
            "description": "Control-plane backend that produced this projection."
          }
        },
        "required": [
          "backend"
        ],
        "type": "object"
      },
      "JobId": {
        "description": "Server-assigned identifier for a job (non-empty).",
        "type": "string"
      },
      "JobInfo": {
        "description": "`GET /jobs/{id}` response: job progress.\n\nContains full detail about a single job, including per-file statuses.\nPolled by the CLI to drive progress bars and by the dashboard for\nlive-updating tables.",
        "properties": {
          "active_lease": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/LeaseRecord",
                "description": "Active lease information when this job is currently claimed by a node."
              },
              {
                "type": "null"
              }
            ]
          },
          "command": {
            "$ref": "#/components/schemas/ReleasedCommand",
            "description": "Batchalign command that was submitted (e.g. \"morphotag\", \"align\")."
          },
          "completed_at": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/MachineTime",
                "description": "When the job reached a terminal state (`MachineTime`)."
              },
              {
                "type": "null"
              }
            ]
          },
          "completed_files": {
            "description": "Number of files that have finished processing (status `Done`).\nInvariant: `0 <= completed_files <= total_files`.",
            "format": "int64",
            "type": "integer"
          },
          "control_plane": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/JobControlPlaneInfo",
                "description": "Per-language-group progress for batched text commands (morphotag,\nServer control-plane metadata for this job when returned by a server."
              },
              {
                "type": "null"
              }
            ]
          },
          "current_file": {
            "description": "Filename currently being processed.  `None` when the job is queued\nor has reached a terminal state.",
            "type": [
              "string",
              "null"
            ]
          },
          "duration_s": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/NonNegativeSeconds",
                "description": "Seconds from submission to completion. `None` while the job is\nstill active."
              },
              {
                "type": "null"
              }
            ]
          },
          "error": {
            "description": "Job-level error message (e.g. worker pool exhaustion, memory gate\ntimeout).  Distinct from per-file errors in `file_statuses`.",
            "type": [
              "string",
              "null"
            ]
          },
          "execution_plan": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/ExecutionPlan",
                "description": "Execution plan describing where and how this job is processed.\nPresent for staged-remote jobs; `None` for local execution."
              },
              {
                "type": "null"
              }
            ]
          },
          "file_statuses": {
            "description": "Per-file processing status, one entry per submitted file.",
            "items": {
              "$ref": "#/components/schemas/FileStatusEntry"
            },
            "type": "array"
          },
          "job_id": {
            "$ref": "#/components/schemas/JobId",
            "description": "Server-assigned UUID (v4) for this job."
          },
          "lang": {
            "$ref": "#/components/schemas/LanguageSpec",
            "description": "Job language: a resolved ISO 639-3 code, `\"auto\"`, `\"per-file\"`, or a\ncode-switched pair such as `\"eng,spa\"`."
          },
          "last_cancelled_at": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/MachineTime",
                "description": "Wall-clock timestamp of the most recent cancel attempt against\nthis job. `None` until at least one cancel arrives. Denormalized\nfrom the `cancellations` audit table for fast list-view rendering\nwithout a JOIN. See `CancellationRecord` for the full audit\nhistory via `GET /jobs/{id}/cancellations`."
              },
              {
                "type": "null"
              }
            ]
          },
          "last_cancelled_host": {
            "description": "Caller-reported host of the most recent cancel attempt.",
            "type": [
              "string",
              "null"
            ]
          },
          "last_cancelled_reason": {
            "description": "Caller-reported reason of the most recent cancel attempt.",
            "type": [
              "string",
              "null"
            ]
          },
          "last_cancelled_source": {
            "description": "Wire-format source string of the most recent cancel attempt\n(`tui`, `api`, `signal`, ...). See `CancelSource`.",
            "type": [
              "string",
              "null"
            ]
          },
          "next_eligible_at": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/MachineTime",
                "description": "When the job is deferred before execution, the earliest time\nat which it should be retried. `None` when the job is immediately\neligible to run."
              },
              {
                "type": "null"
              }
            ]
          },
          "num_workers": {
            "description": "Number of concurrent Python workers used for this job.  Determined\nby the pool at dispatch time.  `None` for queued jobs.",
            "format": "int64",
            "type": [
              "integer",
              "null"
            ]
          },
          "options": {
            "description": "Typed submitted command options captured at job creation time.\n\nThe dashboard uses this to show the original argument/config payload for\ndebugging and operator review without reconstructing a lossy CLI string."
          },
          "source_dir": {
            "description": "Client's original input directory path, used for display in the\ndashboard and CLI output.  May be empty for content-mode submissions.",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/JobStatus",
            "description": "Current lifecycle state of the job."
          },
          "submitted_at": {
            "$ref": "#/components/schemas/MachineTime",
            "description": "When the job was submitted (`MachineTime`: UTC, three fractional\ndigits)."
          },
          "submitted_by": {
            "description": "IP address or Tailscale identifier of the submitting client.\nUsed for conflict detection (same `submitted_by` + filename =\nduplicate).",
            "type": [
              "string",
              "null"
            ]
          },
          "submitted_by_name": {
            "description": "Human-readable hostname of the submitting machine, resolved from\nTailscale or reverse DNS.  For dashboard display only.",
            "type": [
              "string",
              "null"
            ]
          },
          "total_files": {
            "description": "Total number of files in this job (immutable after submission).",
            "format": "int64",
            "type": "integer"
          }
        },
        "required": [
          "job_id",
          "status",
          "command",
          "options",
          "total_files",
          "completed_files",
          "submitted_at"
        ],
        "type": "object"
      },
...
      "JobResultResponse": {
        "description": "`GET /jobs/{id}/results` response: completed job results.",
        "properties": {
          "files": {
            "description": "Per-file results (empty until the job completes).",
            "items": {
              "$ref": "#/components/schemas/FileResult"
            },
            "type": "array"
          },
          "job_id": {
            "$ref": "#/components/schemas/JobId",
            "description": "Unique job identifier."
          },
          "status": {
            "$ref": "#/components/schemas/JobStatus",
            "description": "Terminal status of the job."
          }
        },
        "required": [
          "job_id",
          "status"
        ],
        "type": "object"
      },
      "JobStatus": {
        "description": "Lifecycle states for a job, mirrors `JobStatus` enum.\n\nJobs progress through a linear lifecycle: `Queued -> Running -> {terminal}`.\nTerminal states (`Completed`, `Failed`, `Cancelled`, `Interrupted`) are\npermanent and never transition further.  Only `Cancelled` and `Failed` jobs\ncan be restarted (which resets them to `Queued`).",
        "enum": [
          "queued",
          "running",
          "completed",
          "failed",
          "cancelled",
          "interrupted",
          "writeback_failed"
        ],
        "type": "string"
      },
...
      "LanguageSpec": {
        "description": "Job language: `auto` (the ASR engine detects one language), `per-file` (each file's `@Languages:` header decides), an ISO 639-3 code such as `eng`, or a code-switched pair such as `eng,spa`, primary language first.",
        "examples": [
          "eng",
          "auto",
          "per-file",
          "eng,spa"
        ],
        "type": "string"
      },
      "LeaseRecord": {
        "description": "Lease metadata for a claimed schedulable unit.\n\nIn single-node mode this is coarse-grained and attached at the job level:\nthe local dispatcher claims a queued job for a node, records the lease, and\nlater clears it when the runner exits. Fleet mode will extend the same shape\nwith real cross-node renewal and expiry handling.\n\nA lease always expires strictly after its heartbeat.",
        "properties": {
          "expires_at": {
            "$ref": "#/components/schemas/MachineTime",
            "description": "When the lease expires if not renewed."
          },
          "heartbeat_at": {
            "$ref": "#/components/schemas/MachineTime",
            "description": "When the lease was taken or most recently renewed."
          },
          "leased_by_node": {
            "$ref": "#/components/schemas/NodeId",
            "description": "Identifier of the node that currently owns the lease."
          }
        },
        "required": [
          "leased_by_node",
          "heartbeat_at",
          "expires_at"
        ],
        "type": "object"
      },
      "MachineTime": {
        "description": "RFC 3339 in UTC with exactly three fractional digits, so string order is time order.",
        "examples": [
          "2026-09-28T16:00:00.000Z"
        ],
        "format": "date-time",
        "type": "string"
      },
...
      "NodeId": {
        "description": "Identifier of a server/fleet node.\nEmpty when the node does not report an identity (older server versions).",
        "type": "string"
      },
      "NonNegativeSeconds": {
        "description": "A finite, non-negative number of seconds.",
        "format": "double",
        "minimum": 0,
        "type": "number"
      },
...
      "ReleasedCommand": {
        "description": "Closed released command vocabulary used at all Rust seams.\n\nThis is the single canonical command type. Unknown command strings are\nrejected at deserialization boundaries (HTTP 422, DB recovery skip).",
        "enum": [
          "align",
          "transcribe",
          "transcribe_s",
          "translate",
          "morphotag",
          "coref",
          "utseg",
          "benchmark",
          "opensmile",
          "compare",
          "avqi",
          "diarize",
          "speaker_identify"
        ],
        "type": "string"
      },
...
      "StatusMessageResponse": {
        "description": "Response body for mutating operations that return a status confirmation\n(e.g. cancel, delete, restart).\n\nReturned with HTTP 200 on success. Clients can branch on `status` for\nprogrammatic handling and display `message` to users.",
        "properties": {
          "message": {
            "description": "Human-readable sentence describing the action taken (e.g.\n`\"Job abc123 has been cancelled.\"`).",
            "type": "string"
          },
          "status": {
            "description": "Machine-readable outcome token (e.g. `\"cancelled\"`, `\"deleted\"`,\n`\"restarted\"`). Suitable for switch/match in client code.",
            "type": "string"
          }
        },
        "required": [
          "status",
          "message"
        ],
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "batchalign3: command-line interface, application, server, worker pool, cache",
    "license": {
      "identifier": "BSD-3-Clause",
      "name": "BSD-3-Clause"
    },
    "title": "batchalign",
    "version": "0.8.5"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/jobs/{job_id}": {
      "delete": {
        "description": "Returns 409 if the job is still running -- the caller must cancel it first.\nDeleting a job removes it from the in-memory store and SQLite, and broadcasts\na `JobDeleted` event to connected WebSocket/SSE clients.",
        "operationId": "delete_job",
        "parameters": [
          {
            "description": "Job identifier",
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatusMessageResponse"
                }
              }
            },
            "description": "Job deleted"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Job not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Job still running"
          }
        },
        "summary": "Permanently remove a terminal job and its associated state.",
        "tags": [
          "jobs"
        ]
      },
      "get": {
        "description": "This is the primary polling endpoint for the CLI: the client calls it\nrepeatedly until `status` reaches a terminal value.",
        "operationId": "get_job",
        "parameters": [
          {
            "description": "Job identifier",
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobInfo"
                }
              }
            },
            "description": "Job detail"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Job not found"
          }
        },
        "summary": "Return full detail for a single job, including per-file statuses.",
        "tags": [
          "jobs"
        ]
      }
    },
...
    "/jobs/{job_id}/results": {
      "get": {
        "description": "Blocks until the job has completed, failed, or been interrupted -- returns\n409 for jobs that are still running or queued. The response includes the\nfull staged content of each successful output file. Paths-mode jobs keep\nthe execution-host write, but also preserve a staged copy for download.\nCancelled jobs are intentionally excluded because they may have incomplete\nresults.",
        "operationId": "get_results",
        "parameters": [
          {
            "description": "Job identifier",
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobResultResponse"
                }
              }
            },
            "description": "Results for terminal jobs"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Job not found"
          }
        },
        "summary": "Retrieve all output files for a terminal job.",
        "tags": [
          "jobs"
        ]
      }
    },
    "/jobs/{job_id}/results/{*filename}": {
      "get": {
        "description": "Waits for the individual file to reach a terminal status, not the whole job,\nso clients can fetch results incrementally as files finish. Supports\ncommand-policy matching (e.g., requesting `sample.wav` finds `sample.cha`,\nand speaker identification maps `sample.cha` to its evidence JSON).",
        "operationId": "get_single_result",
        "parameters": [
          {
            "description": "Job identifier",
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Requested output filename",
            "in": "path",
            "name": "filename",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FileResult"
                }
              }
            },
            "description": "One file result"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "File not found"
          }
        },
        "summary": "Retrieve the output for a single file within a job.",
        "tags": [
          "jobs"
        ]
      }
    },
...
  },
...
}